refactor: rename package from 'domain' to 'entity' and update navbar behavior
