+++ src/main/java/vn/demo/nike/features/catalog/cart/entity/CartItem.java
@@ -4,7 +4,7 @@ import jakarta.persistence.*;
 import lombok.*;
 import vn.demo.nike.features.catalog.product.entity.ProductVariant;
 import vn.demo.nike.features.identity.user.entity.User;
-import vn.demo.nike.shared.domain.BaseEntity;
+import vn.demo.nike.shared.entity.BaseEntity;
 
 @Entity
 @Table(name = "cart_items", uniqueConstraints = {
