+++ src/main/java/vn/demo/nike/features/catalog/category/entity/Category.java
@@ -2,7 +2,7 @@ package vn.demo.nike.features.catalog.category.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
-import vn.demo.nike.shared.domain.BaseEntity;
+import vn.demo.nike.shared.entity.BaseEntity;
 import vn.demo.nike.features.catalog.product.entity.Product;
 
 import java.util.List;
