+++ src/main/java/vn/demo/nike/features/catalog/product/entity/Product.java
@@ -2,8 +2,7 @@ package vn.demo.nike.features.catalog.product.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
-import lombok.experimental.SuperBuilder;
-import vn.demo.nike.shared.domain.BaseEntity;
+import vn.demo.nike.shared.entity.BaseEntity;
 import vn.demo.nike.features.catalog.category.entity.Category;
 import vn.demo.nike.features.catalog.product.enums.ProductStatus;
