+++ src/main/java/vn/demo/nike/shared/entity/BaseEntity.java
@@ -1,4 +1,4 @@
-package vn.demo.nike.shared.domain;
+package vn.demo.nike.shared.entity;
 
 import java.time.LocalDateTime;
