+++ src/main/resources/static/css/customer/product-list/product-list-responsive.css
@@ -42,7 +42,7 @@
 }
 
 /* Navigation Responsive - Large Tablets and Below */
-@media (max-width: 1180px) {
+@media (max-width: 992px) {
     .nav-center {
         display: none;
     }
