+++ src/main/resources/static/js/customer/components/header.js
@@ -1,4 +1,17 @@
 (function () {
+  function syncHeaderOffset() {
+    const header = document.getElementById("header");
+    const headerTop = header ? header.querySelector(".header-top") : null;
+    const navbar = header ? header.querySelector(".navbar") : null;
+    const isMobileTopHidden = window.innerWidth <= 768;
+    const headerTopHeight = headerTop && !isMobileTopHidden ? headerTop.offsetHeight : 0;
+    const navHeight = navbar ? navbar.offsetHeight : 0;
+
+    document.documentElement.style.setProperty("--header-top-height", headerTopHeight + "px");
+    document.documentElement.style.setProperty("--header-nav-height", navHeight + "px");
+    document.documentElement.style.setProperty("--header-visible-offset", (headerTopHeight + navHeight) + "px");
+  }
+
   function bindGlobalSearchFallback() {
     const form = document.getElementById("globalSearchForm");
     if (!form) {
@@ -53,9 +66,19 @@
   }
 
   function init() {
+    syncHeaderOffset();
     bindGlobalSearchFallback();
     bindHeaderActions();
     bindPageShowReload();
+    window.addEventListener("resize", syncHeaderOffset);
+
+    if (typeof ResizeObserver === "function") {
+      const header = document.getElementById("header");
+      if (header) {
+        const resizeObserver = new ResizeObserver(syncHeaderOffset);
+        resizeObserver.observe(header);
+      }
+    }
   }
 
   if (document.readyState === "loading") {
